This is an initial merge of the current Eucalyptatron POC and the Regalbot code currently found on the dev staging environment
This commit shows some reorganization of the file structure, the start
of the migration from Typescript to Flow Types and the mere beginnings
a refactoring of the RB entry points into the new Claudia.js ones.

But running the code itself... that is a minefield currently. 🚒
Do not attempt. :skull: